+++ pyvkfft/accuracy.py
@@ -66,7 +66,7 @@ except ImportError:
 gpu_ctx_dic = {}
 
 
-def init_ctx(backend, gpu_name=None, verbose=False):
+def init_ctx(backend, gpu_name=None, opencl_platform=None, verbose=False):
     if backend in gpu_ctx_dic:
         return
     if backend == "pycuda":
@@ -96,6 +96,9 @@ def init_ctx(backend, gpu_name=None, verbose=False):
         for p in cl.get_platforms():
             if d is not None:
                 break
+            if opencl_platform is not None:
+                if opencl_platform.lower() not in p.name.lower():
+                    continue
             for d0 in p.get_devices():
                 if d0.type & cl.device_type.GPU:
                     if gpu_name is not None:
@@ -106,10 +109,8 @@ def init_ctx(backend, gpu_name=None, verbose=False):
                 if d is not None:
                     break
         if d is None:
-            if gpu_name is not None:
-                raise RuntimeError("Selected backend is pyopencl, but no device found (name=%s)" % gpu_name)
-            else:
-                raise RuntimeError("Selected backend is pyopencl, but no device found")
+            raise RuntimeError("Selected backend is pyopencl, but no device found (name=%s, platform=%s)" %
+                               (gpu_name, opencl_platform))
         cl_ctx = cl.Context([d])
         cq = cl.CommandQueue(cl_ctx)
         gpu_ctx_dic["pyopencl"] = d, cl_ctx, cq, 'cl_khr_fp64' in cq.device.extensions
@@ -153,13 +154,13 @@ def li(a, b):
 
 
 def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c=False, dct=False,
-                  gpu_name=None, stream=None, queue=None, return_array=False, init_array=None, verbose=False,
-                  colour_output=False, ref_long_double=True):
+                  gpu_name=None, opencl_platform=None, stream=None, queue=None, return_array=False,
+                  init_array=None, verbose=False, colour_output=False, ref_long_double=True, order='C'):
     """
     Measure the
     :param backend: either 'pyopencl', 'pycuda' or 'cupy'
     :param shape: the shape of the array to test. If this is an inplace r2c, the
-        x-axis length must be even, and two extra values will be appended along x,
+        fast-axis length must be even, and two extra values will be appended along x,
         so the actual transform shape is the one supplied
     :param ndim: the number of FFT dimensions. Can be None if axes is given
     :param axes: the transform axes. Supersedes ndim
@@ -175,6 +176,8 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         tested (native scipy normalisation).
     :param gpu_name: the name of the gpu to use. If None, the first available
         for the backend will be used.
+    :param opencl_platform: the name of the OpenCL platform to use. If None, the first available
+        will be used.
     :param stream: the cuda stream to use, or None
     :param queue: the opencl queue to use (mandatory for the 'pyopencl' backend)
     :param return_array: if True, will return the generated random array so it can be
@@ -187,6 +190,10 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
     :param colour_output: if True, use some colour to tag the quality of the accuracy
     :param ref_long_double: if True and scipy is available, long double precision
         will be used for the reference transform. Otherwise, this is ignored.
+    :param order: either 'C' (default C-contiguous) or 'F' to test a different
+        stride. Note that for the latter, a 3D transform on a 4D array will not
+        be supported as the last transform axis would be on the 4th dimension
+        (once ordered by stride).
     :return: a dictionary with (l2_fft, li_fft, l2_ifft, li_ifft, tol, dt_array,
         dt_app, dt_fft, dt_ifft, src_unchanged_fft, src_unchanged_ifft, tol_test, str),
         with the L2 and Linf normalised norms comparing pyvkfft's result with either
@@ -204,6 +211,7 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         All input parameters are also returned as key/values, except stream, queue,
         return_array, ini_array and verbose.
     """
+    ndims = len(shape)
     if backend == "cupy" and has_cupy:
         mempool = cp.get_default_memory_pool()
         if mempool is not None:  # Is that test necessary ?
@@ -211,7 +219,7 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
             # N parallel process so memory management  must be done manually
             mempool.free_all_blocks()
     t0 = timeit.default_timer()
-    init_ctx(backend, gpu_name=gpu_name, verbose=False)
+    init_ctx(backend, gpu_name=gpu_name, opencl_platform=opencl_platform, verbose=False)
     if backend == "pyopencl" and queue is None:
         queue = gpu_ctx_dic["pyopencl"][2]
     shape0 = shape
@@ -231,10 +239,16 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
             # Add two extra columns in the source array
             # so the transform has the desired shape
             shape = list(shape)
-            shape[-1] += 2
+            if order == 'C':
+                shape[-1] += 2
+            else:
+                shape[0] += 2
         else:
             shapec = list(shape)
-            shapec[-1] = shapec[-1] // 2 + 1
+            if order == 'C':
+                shapec[-1] = shapec[-1] // 2 + 1
+            else:
+                shapec[0] = shapec[-1] // 2 + 1
             shapec = tuple(shapec)
     else:
         shapec = tuple(shape)
@@ -244,7 +258,10 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         if r2c:
             if inplace:
                 d0 = np.empty(shape, dtype=dtypef)
-                d0[..., :-2] = init_array
+                if order == 'C':
+                    d0[..., :-2] = init_array
+                else:
+                    d0[:-2] = init_array
             else:
                 d0 = init_array.astype(dtypef)
         elif dct:
@@ -257,32 +274,55 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         else:
             d0 = (np.random.uniform(-0.5, 0.5, shape) + 1j * np.random.uniform(-0.5, 0.5, shape)).astype(dtype)
 
+    if order != 'C':
+        d0 = np.asarray(d0, order=order)
+
     t1 = timeit.default_timer()
 
     if 'opencl' in backend:
         app = clVkFFTApp(d0.shape, d0.dtype, queue, ndim=ndim, norm=norm,
-                         axes=axes, useLUT=use_lut, inplace=inplace, r2c=r2c, dct=dct)
+                         axes=axes, useLUT=use_lut, inplace=inplace, r2c=r2c, dct=dct, strides=d0.strides)
         t2 = timeit.default_timer()
         d_gpu = cla.to_device(queue, d0)
+        empty_like = cla.empty_like
     else:
         if backend == "pycuda":
             to_gpu = cua.to_gpu
+            empty_like = cua.empty_like
         else:
             to_gpu = cp.array
+            empty_like = cp.empty_like
 
         app = cuVkFFTApp(d0.shape, d0.dtype, ndim=ndim, norm=norm, axes=axes,
-                         useLUT=use_lut, inplace=inplace, r2c=r2c, dct=dct, stream=stream)
+                         useLUT=use_lut, inplace=inplace, r2c=r2c, dct=dct, strides=d0.strides, stream=stream)
         t2 = timeit.default_timer()
         d_gpu = to_gpu(d0)
 
     if axes is None:
-        axes_numpy = list(range(len(shape)))[-ndim:]
+        axes_numpy = list(range(ndims))[-ndim:]
     else:
-        axes_numpy = axes
+        # Make sure axes indices are >0
+        axes_numpy = [ax if ax >= 0 else ndims + ax for ax in axes]
+
+    # Need the fast axis for R2C (last for 'C' order, first for 'F')
+    fast_axis = np.argmin(d0.strides)
+
+    if r2c:
+        if fast_axis not in axes_numpy:
+            raise RuntimeError("The fast axis must be transformed for R2C")
+
+        # For R2C, we need the same fast axis as on the GPU, or the
+        # half-hermitian result won't look the same
+        if fast_axis != axes_numpy[-1]:
+            axes_numpy.remove(fast_axis)
+            axes_numpy.append(fast_axis)
+        # if order != 'C':
+        #     print("R2C", ndims, ndim, shape, axes, axes_numpy, fast_axis, inplace)
+
     # base FFT scale for numpy (not used for DCT)
     s = np.sqrt(np.prod([d0.shape[i] for i in axes_numpy]))
     if r2c and inplace:
-        s = np.sqrt(s ** 2 / d0.shape[-1] * (d0.shape[-1] - 2))
+        s = np.sqrt(s ** 2 / d0.shape[fast_axis] * (d0.shape[fast_axis] - 2))
 
     # Tolerance estimated from accuracy notebook
     if dtype in (np.complex64, np.float32):
@@ -301,13 +341,14 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
     else:
         if r2c:
             if backend == "pyopencl":
-                d1_gpu = cla.empty(queue, shapec, dtype=dtype)
+                d1_gpu = cla.empty(queue, shapec, dtype=dtype, order=order)
             elif backend == "pycuda":
-                d1_gpu = cua.empty(shapec, dtype=dtype)
+                d1_gpu = cua.empty(shapec, dtype=dtype, order=order)
             elif backend == "cupy":
-                d1_gpu = cp.empty(shapec, dtype=dtype)
+                d1_gpu = cp.empty(shapec, dtype=dtype, order=order)
         else:
-            d1_gpu = d_gpu.copy()
+            # Note that pycuda's gpuarray.copy (as of 2022.2.2) does not copy strides
+            d1_gpu = empty_like(d_gpu)
 
     if has_scipy and ref_long_double:
         # Use long double precision
@@ -324,7 +365,8 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
 
     if r2c:
         if inplace:
-            d = rfftn(d0n[..., :-2], axes=axes_numpy) / s
+            # Need to cut the fastest axis by 2
+            d = rfftn(np.take(d0n, range(d0n.shape[fast_axis] - 2), axis=fast_axis), axes=axes_numpy) / s
         else:
             d = rfftn(d0n, axes=axes_numpy) / s
     elif dct:
@@ -348,12 +390,20 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         t = "C2C"
     if r2c and inplace:
         tmp = list(d0.shape)
-        tmp[-1] -= 2
-        shstr = str(tuple(tmp)).replace(" ", "")
-        if ",)" in shstr:
-            shstr = shstr.replace(",)", "+2)")
+        if order == 'C':
+            tmp[-1] -= 2
+            shstr = str(tuple(tmp)).replace(" ", "")
+            if ",)" in shstr:
+                shstr = shstr.replace(",)", "+2)")
+            else:
+                shstr = shstr.replace(")", "+2)")
         else:
-            shstr = shstr.replace(")", "+2)")
+            tmp[0] -= 2
+            if len(tmp) == 1:
+                shstr = "(%d+2)," % tmp[0]
+            else:
+                shstr = "(%d+2," % tmp[0]
+                shstr += str(tuple(tmp[1:])).replace(" ", "").replace(",)", ")")[1:]
     else:
         shstr = str(d0.shape).replace(" ", "")
     shax = str(axes).replace(" ", "")
@@ -364,9 +414,9 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
         stol = "%6.2e < %6.2e (%5.3f)" % (ni, tol, ni / tol)
 
     verb_out = "%8s %4s %14s axes=%10s ndim=%4s %10s lut=%4s inplace=%d " \
-               " norm=%4s %5s: n2=%6.2e ninf=%s %d" % \
+               " norm=%4s %s %5s: n2=%6.2e ninf=%s %d" % \
                (backend, t, shstr, shax, str(ndim), str(d0.dtype),
-                str(use_lut), int(inplace), str(norm), "FFT", n2, stol, src_unchanged_fft)
+                str(use_lut), int(inplace), str(norm), order, "FFT", n2, stol, src_unchanged_fft)
 
     t3 = timeit.default_timer()
 
@@ -387,7 +437,11 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
             d0n = d0.astype(np.clongdouble)
         else:
             d0n = d0
-
+        if (np.argmin(d0.strides) != d0.ndim - 1) and order == 'C':
+            # np.fft.rfftn can change the fast axis
+            d0 = np.asarray(d0, order='C')
+    if order != 'C':
+        d0 = np.asarray(d0, order=order)
     if 'opencl' in backend:
         d_gpu = cla.to_device(queue, d0)
     else:
@@ -398,13 +452,13 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
     else:
         if r2c:
             if backend == "pyopencl":
-                d1_gpu = cla.empty(queue, shape, dtype=dtypef)
+                d1_gpu = cla.empty(queue, shape, dtype=dtypef, order=order)
             elif backend == "pycuda":
-                d1_gpu = cua.empty(shape, dtype=dtypef)
+                d1_gpu = cua.empty(shape, dtype=dtypef, order=order)
             elif backend == "cupy":
-                d1_gpu = cp.empty(shape, dtype=dtypef)
+                d1_gpu = cp.empty(shape, dtype=dtypef, order=order)
         else:
-            d1_gpu = d_gpu.copy()
+            d1_gpu = empty_like(d_gpu)
 
     d1_gpu = app.ifft(d_gpu, d1_gpu)
     if not dct:
@@ -424,7 +478,8 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
             assert d1_gpu.dtype == dtype, "The array type is incorrect after an inplace iFFT"
 
     if r2c and inplace:
-        n2i, nii = l2(d, d1_gpu.get()[..., :-2]), li(d, d1_gpu.get()[..., :-2])
+        tmp = np.take(d1_gpu.get(), range(d1_gpu.shape[fast_axis] - 2), axis=fast_axis)
+        n2i, nii = l2(d, tmp), li(d, tmp)
     else:
         n2i, nii = l2(d, d1_gpu.get()), li(d, d1_gpu.get())
 
@@ -461,7 +516,7 @@ def test_accuracy(backend, shape, ndim, axes, dtype, inplace, norm, use_lut, r2c
            "dt_fft": t3 - t2, "dt_ifft": t4 - t3, "src_unchanged_fft": src_unchanged_fft,
            "src_unchanged_ifft": src_unchanged_ifft, "tol_test": max(ni, nii) < tol, "str": verb_out,
            "backend": backend, "shape": shape0, "ndim": ndim, "axes": axes, "dtype": dtype0, "inplace": inplace,
-           "norm": norm, "use_lut": use_lut, "r2c": r2c, "dct": dct, "gpu_name": gpu_name}
+           "norm": norm, "use_lut": use_lut, "r2c": r2c, "dct": dct, "gpu_name": gpu_name, "order": order}
 
     if return_array:
         res["d0"] = d0