+++ pyvkfft/test/test_fft.py
@@ -69,6 +69,18 @@ class TestFFT(unittest.TestCase):
     nproc = 1
     verbose = False
     colour = False
+    opencl_platform = None
+    vbackend = None
+
+    def setUp(self) -> None:
+        if self.vbackend is None:
+            self.vbackend = []
+            if has_pycuda:
+                self.vbackend.append("pycuda")
+            if has_cupy:
+                self.vbackend.append("cupy")
+            if has_pyopencl:
+                self.vbackend.append("pyopencl")
 
     def test_backend(self):
         self.assertTrue(has_pycuda or has_pyopencl or has_cupy,
@@ -77,17 +89,9 @@ class TestFFT(unittest.TestCase):
     @unittest.skipIf(not (has_pycuda or has_cupy or has_pyopencl), "No OpenCL/CUDA backend is available")
     def test_simple_fft(self):
         """Test the simple fft API"""
-        vbackend = []
-        if has_pycuda:
-            vbackend.append("pycuda")
-        if has_cupy:
-            vbackend.append("cupy")
-        if has_pyopencl:
-            vbackend.append("pyopencl")
-
-        for backend in vbackend:
+        for backend in self.vbackend:
             with self.subTest(backend=backend):
-                init_ctx(backend, gpu_name=self.gpu, verbose=False)
+                init_ctx(backend, gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
                 if backend == "pycuda":
                     dc = cua.to_gpu(ascent().astype(np.complex64))
                     dr = cua.to_gpu(ascent().astype(np.float32))
@@ -163,7 +167,7 @@ class TestFFT(unittest.TestCase):
         ct = 0
         vkwargs = []
         for backend in vbackend:
-            init_ctx(backend, gpu_name=self.gpu, verbose=False)
+            init_ctx(backend, gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
             cq = gpu_ctx_dic["pyopencl"][2] if backend == "pyopencl" else None
             for n in vn:
                 for dims in range(1, dims_max + 1):
@@ -201,7 +205,7 @@ class TestFFT(unittest.TestCase):
                                                       + 1j * np.random.uniform(-0.5, 0.5, sh)).astype(dtype)
                                         if vlut == "auto":
                                             if dtype in (np.float64, np.complex128):
-                                                # By default LUT is enabled for complex128, no need to test twice
+                                                # By default, LUT is enabled for complex128, no need to test twice
                                                 tmp = [None]
                                             else:
                                                 tmp = [None, True]
@@ -210,50 +214,71 @@ class TestFFT(unittest.TestCase):
                                         for use_lut in tmp:
                                             for inplace in vinplace:
                                                 for norm in vnorm:
-                                                    with self.subTest(backend=backend, n=n, dims=dims, ndim=ndim,
-                                                                      axes=axes, dtype=np.dtype(dtype), norm=norm,
-                                                                      use_lut=use_lut, inplace=inplace,
-                                                                      r2c=r2c, dct=dct):
-                                                        ct += 1
-                                                        if not dry_run:
-                                                            res = test_accuracy(backend, sh, ndim, axes, dtype, inplace,
-                                                                                norm, use_lut, r2c=r2c, dct=dct,
-                                                                                gpu_name=self.gpu,
-                                                                                stream=None, queue=cq,
-                                                                                return_array=False, init_array=d0,
-                                                                                verbose=verbose)
-                                                            npr = primes(n)
-                                                            ni, n2 = res["ni"], res["n2"]
-                                                            nii, n2i = res["nii"], res["n2i"]
-                                                            tol = res["tol"]
-                                                            src1 = res["src_unchanged_fft"]
-                                                            src2 = res["src_unchanged_ifft"]
-                                                            self.assertTrue(ni < tol, "Accuracy mismatch after FFT, "
-                                                                                      "n2=%8e ni=%8e>%8e" %
-                                                                            (n2, ni, tol))
-                                                            self.assertTrue(nii < tol, "Accuracy mismatch after iFFT, "
-                                                                                       "n2=%8e ni=%8e>%8e" %
-                                                                            (n2, nii, tol))
-                                                            if not inplace:
-                                                                self.assertTrue(src1, "The source array was modified "
-                                                                                      "during the FFT")
-                                                                nmaxr2c1d = 3072 * (1 + int(
-                                                                    dtype in (np.float32, np.complex64)))
-                                                                if not r2c or (ndim == 1 and max(npr) <= 13) \
-                                                                        and n < nmaxr2c1d:
-                                                                    self.assertTrue(src2,
+                                                    vorder = ['C', 'F']
+                                                    if dims == 1 or dims > 3 or dct:
+                                                        vorder = ['C']
+                                                    if r2c:
+                                                        if ndim is not None:
+                                                            if dims != ndim:
+                                                                vorder = ['C']
+                                                        if axes is not None:
+                                                            # TODO : also test F-order when ndim<dims but fast axis
+                                                            #  is transformed
+                                                            if len(axes) != dims:
+                                                                vorder = ['C']
+                                                    for order in vorder:
+                                                        with self.subTest(backend=backend, n=n, dims=dims, ndim=ndim,
+                                                                          axes=axes, dtype=np.dtype(dtype), norm=norm,
+                                                                          use_lut=use_lut, inplace=inplace,
+                                                                          r2c=r2c, dct=dct, order=order):
+                                                            ct += 1
+                                                            if not dry_run:
+                                                                res = \
+                                                                    test_accuracy(backend, sh, ndim, axes, dtype,
+                                                                                  inplace,
+                                                                                  norm, use_lut, r2c=r2c, dct=dct,
+                                                                                  gpu_name=self.gpu,
+                                                                                  opencl_platform=self.opencl_platform,
+                                                                                  stream=None, queue=cq,
+                                                                                  return_array=False, init_array=d0,
+                                                                                  verbose=verbose, order=order)
+                                                                npr = primes(n)
+                                                                ni, n2 = res["ni"], res["n2"]
+                                                                nii, n2i = res["nii"], res["n2i"]
+                                                                tol = res["tol"]
+                                                                src1 = res["src_unchanged_fft"]
+                                                                src2 = res["src_unchanged_ifft"]
+                                                                self.assertTrue(ni < tol,
+                                                                                "Accuracy mismatch after FFT, "
+                                                                                "n2=%8e ni=%8e>%8e" %
+                                                                                (n2, ni, tol))
+                                                                self.assertTrue(nii < tol,
+                                                                                "Accuracy mismatch after iFFT, "
+                                                                                "n2=%8e ni=%8e>%8e" %
+                                                                                (n2, nii, tol))
+                                                                if not inplace:
+                                                                    self.assertTrue(src1,
                                                                                     "The source array was modified "
-                                                                                    "during the iFFT")
-                                                        else:
-                                                            kwargs = {"backend": backend, "shape": sh,
-                                                                      "ndim": ndim, "axes": axes,
-                                                                      "dtype": dtype, "inplace": inplace,
-                                                                      "norm": norm, "use_lut": use_lut,
-                                                                      "r2c": r2c, "dct": dct,
-                                                                      "gpu_name": self.gpu, "stream": None,
-                                                                      "verbose": False,
-                                                                      "colour_output": self.colour}
-                                                            vkwargs.append(kwargs)
+                                                                                    "during the FFT")
+                                                                    nmaxr2c1d = 3072 * (1 + int(
+                                                                        dtype in (np.float32, np.complex64)))
+                                                                    if not r2c or (ndim == 1 and max(npr) <= 13) \
+                                                                            and n < nmaxr2c1d:
+                                                                        self.assertTrue(src2,
+                                                                                        "The source array was modified "
+                                                                                        "during the iFFT")
+                                                            else:
+                                                                kwargs = {"backend": backend, "shape": sh,
+                                                                          "ndim": ndim, "axes": axes,
+                                                                          "dtype": dtype, "inplace": inplace,
+                                                                          "norm": norm, "use_lut": use_lut,
+                                                                          "r2c": r2c, "dct": dct,
+                                                                          "gpu_name": self.gpu,
+                                                                          "opencl_platform": self.opencl_platform,
+                                                                          "stream": None,
+                                                                          "verbose": False, "order": order,
+                                                                          "colour_output": self.colour}
+                                                                vkwargs.append(kwargs)
 
         return ct, vkwargs
 
@@ -263,7 +288,7 @@ class TestFFT(unittest.TestCase):
             for res in pool.imap(test_accuracy_kwargs, vkwargs):
                 with self.subTest(backend=res['backend'], n=max(res['shape']), ndim=res['ndim'],
                                   dtype=np.dtype(res['dtype']), norm=res['norm'], use_lut=res['use_lut'],
-                                  inplace=res['inplace'], r2c=res['r2c'], dct=res['dct']):
+                                  inplace=res['inplace'], r2c=res['r2c'], dct=res['dct'], order=res['order']):
                     n = max(res['shape'])
                     npr = primes(n)
                     ni, n2 = res["ni"], res["n2"]
@@ -286,15 +311,8 @@ class TestFFT(unittest.TestCase):
     @unittest.skipIf(not (has_pycuda or has_cupy or has_pyopencl), "No OpenCL/CUDA backend is available")
     def test_c2c(self):
         """Run C2C tests"""
-        vbackend = []
-        if has_pycuda:
-            vbackend.append("pycuda")
-        if has_cupy:
-            vbackend.append("cupy")
-        if has_pyopencl:
-            vbackend.append("pyopencl")
-        for backend in vbackend:
-            init_ctx(backend, gpu_name=self.gpu, verbose=False)
+        for backend in self.vbackend:
+            init_ctx(backend, gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
             has_cl_fp64 = gpu_ctx_dic["pyopencl"][3] if backend == "pyopencl" else True
             ct = 0
             vkwargs = []
@@ -319,15 +337,8 @@ class TestFFT(unittest.TestCase):
     @unittest.skipIf(not (has_pycuda or has_cupy or has_pyopencl), "No OpenCL/CUDA backend is available")
     def test_r2c(self):
         """Run R2C tests"""
-        vbackend = []
-        if has_pycuda:
-            vbackend.append("pycuda")
-        if has_cupy:
-            vbackend.append("cupy")
-        if has_pyopencl:
-            vbackend.append("pyopencl")
-        for backend in vbackend:
-            init_ctx(backend, gpu_name=self.gpu, verbose=False)
+        for backend in self.vbackend:
+            init_ctx(backend, gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
             has_cl_fp64 = gpu_ctx_dic["pyopencl"][3] if backend == "pyopencl" else True
             ct = 0
             vkwargs = []
@@ -353,15 +364,8 @@ class TestFFT(unittest.TestCase):
     @unittest.skipIf(not has_dct_ref, "scipy and pyfftw are not available - cannot test DCT")
     def test_dct(self):
         """Run DCT tests"""
-        vbackend = []
-        if has_pycuda:
-            vbackend.append("pycuda")
-        if has_cupy:
-            vbackend.append("cupy")
-        if has_pyopencl:
-            vbackend.append("pyopencl")
-        for backend in vbackend:
-            init_ctx(backend, gpu_name=self.gpu, verbose=False)
+        for backend in self.vbackend:
+            init_ctx(backend, gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
             has_cl_fp64 = gpu_ctx_dic["pyopencl"][3] if backend == "pyopencl" else True
             ct = 0
             vkwargs = []
@@ -387,7 +391,7 @@ class TestFFT(unittest.TestCase):
         """
         for dtype in (np.complex64, np.complex128):
             with self.subTest(dtype=np.dtype(dtype)):
-                init_ctx("pycuda", gpu_name=self.gpu, verbose=False)
+                init_ctx("pycuda", gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
                 if dtype == np.complex64:
                     rtol = 1e-6
                 else:
@@ -467,6 +471,7 @@ class TestFFTSystematic(unittest.TestCase):
     gpu = None
     inplace = False
     lut = False
+    fstride = False
     max_pow = None
     max_nb_tests = 1000
     nb_test = 0  # Number of tests actually run
@@ -475,6 +480,7 @@ class TestFFTSystematic(unittest.TestCase):
     # t.ndims = args.ndims
     norm = 1
     nproc = 1
+    opencl_platform = None
     r2c = False
     radix = None
     range = 2, 128
@@ -496,7 +502,7 @@ class TestFFTSystematic(unittest.TestCase):
                 self.vbackend.append("cupy")
             if has_pyopencl:
                 self.vbackend.append("pyopencl")
-                init_ctx("pyopencl", gpu_name=self.gpu, verbose=False)
+                init_ctx("pyopencl", gpu_name=self.gpu, opencl_platform=self.opencl_platform, verbose=False)
                 self.cq, self.has_cl_fp64 = gpu_ctx_dic["pyopencl"][2:]
         self.assertTrue(not self.bluestein or self.radix is None, "Cannot select both Bluestein and radix")
         if not self.bluestein and self.radix is None:
@@ -538,8 +544,10 @@ class TestFFTSystematic(unittest.TestCase):
             for s in self.vshape:
                 kwargs = {"backend": backend, "shape": s, "ndim": len(s), "axes": self.axes,
                           "dtype": self.dtype, "inplace": self.inplace, "norm": self.norm, "use_lut": self.lut,
-                          "r2c": self.r2c, "dct": self.dct, "gpu_name": self.gpu, "stream": None, "verbose": False,
-                          "colour_output": self.colour, "ref_long_double": self.ref_long_double}
+                          "r2c": self.r2c, "dct": self.dct, "gpu_name": self.gpu,
+                          "opencl_platform": self.opencl_platform, "stream": None, "verbose": False,
+                          "colour_output": self.colour, "ref_long_double": self.ref_long_double,
+                          "order": 'F' if self.fstride else 'C'}
                 vkwargs.append(kwargs)
         if self.db is not None:
             # TODO secure the db with a context 'with'
@@ -547,7 +555,7 @@ class TestFFTSystematic(unittest.TestCase):
             dbc = db.cursor()
             dbc.execute('CREATE TABLE IF NOT EXISTS pyvkfft_test (epoch int, hostname int,'
                         'backend text, language text, transform text, axes text, array_shape text,'
-                        'ndims int, ndim int, precision int, inplace int, norm int, lut int,'
+                        'ndims int, ndim int, precision int, inplace int, norm int, lut int, fstride int,'
                         'n int, n2_fft float, n2_ifft float, ni_fft float, ni_ifft float, tolerance float,'
                         'dt_app float, dt_fft float, dt_ifft float, src_unchanged_fft int, src_unchanged_ifft int,'
                         'gpu_name text, success int, error int, vkfft_error_code int)')
@@ -587,7 +595,7 @@ class TestFFTSystematic(unittest.TestCase):
                     # as e.g. "float32" instead of "<class 'numpy.float32'>"
                     with self.subTest(backend=backend, shape=sh, ndim=ndim,
                                       dtype=np.dtype(self.dtype), norm=self.norm, use_lut=self.lut,
-                                      inplace=self.inplace, r2c=self.r2c, dct=self.dct):
+                                      inplace=self.inplace, r2c=self.r2c, dct=self.dct, fstride=self.fstride):
                         if self.serial:
                             res = test_accuracy_kwargs(v)
                         else:
@@ -624,12 +632,12 @@ class TestFFTSystematic(unittest.TestCase):
                                 succ = False
                         if self.db is not None:
                             dbc.execute('INSERT INTO pyvkfft_test VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,'
-                                        '?,?,?,?,?,?,?,?,?,?,?,?,?)',
+                                        '?,?,?,?,?,?,?,?,?,?,?,?,?,?)',
                                         (time.time(), hostname, backend, lang, transform,
                                          str(res['axes']).encode('ascii'), str(res['shape']).encode('ascii'),
                                          len(res['shape']), ndim, np.dtype(self.dtype).itemsize,
-                                         self.inplace, self.norm, self.lut, int(max(res['shape'])), float(n2),
-                                         float(n2i),
+                                         self.inplace, self.norm, self.lut, self.fstride,
+                                         int(max(res['shape'])), float(n2), float(n2i),
                                          float(ni), float(nii), float(tol), res["dt_app"], res["dt_fft"],
                                          res["dt_ifft"],
                                          int(src1), int(src2), res["gpu_name"].encode('ascii'), int(succ), 0, 0))